feat: Add home lite and pro
