+++ src/routes/routes.ts
@@ -6,7 +6,7 @@ import { Transfer } from '../views/Transfer'
 import { Profile } from '../views/Profile'
 import { ImagePreview } from '../views/Profile/ImagePreview'
 import { TakePhoto } from '../views/Profile/TakePhoto'
-import { Explore } from '../views/Explore'
+import { Explore } from '../views/Explore/index'
 import { Help } from '../views/Help'
 import { Unipass } from '../views/Unipass'
 import { Apps } from '../views/Apps'
